requirements.txt
astroid==2.3.3
colorama==0.4.3
cursor==1.3.4
halo==0.0.28
isort==4.3.21
lazy-object-proxy==1.4.3
log-symbols==0.0.14
mccabe==0.6.1
pkg-resources==0.0.0
pycryptodome==3.9.4
pylint==2.4.4
six==1.13.0
spinners==0.0.23
termcolor==1.1.0
typed-ast==1.4.0
wrapt==1.11.2
